scoring_engine/requirements.txt
flask>=3.0
dnspython>=2.6
